@@ -7,6 +7,8 @@
   "dependencies": {
     "soucouyant": "^0.1.3"
   },
+  "type": "module",
+  "cjs": true,
   "devDependencies": {
     "chalk": "^3.0.0",
     "chokidar": "^3.3.0",
@@ -17,13 +19,13 @@
     "ws": "^7.2.0"
   },
   "engines": {
-    "node": ">=13"
+    "node": ">=13.2.0"
   },
   "scripts": {
-    "start": "node setup/server.js",
-    "markup": "yarn start --preview-markup",
-    "distribute": "yarn clean && yarn start --distribute",
-    "clean": "rm -rf ./dist",
+    "start": "node --experimental-json-modules setup/server.js",
+    "markup": "npm run start preview-markup",
+    "dist": "npm run delete-dist && npm run start distribute",
+    "delete-dist": "rm -rf ./dist",
     "validate": ""
   }
 }
